SES-19-David refactoring code
- deleted cart page
- moved fetch method to api.ts
- moved route handlers to /api folder
