+++ src/app/api.ts
@@ -1,4 +1,6 @@
 import { BASE_URL } from "../constants";
+import { cookies } from "next/headers";
+import { AUTH_COOKIE_KEY } from "../constants";
 
 export async function getUsers() {
     const response = await fetch(BASE_URL + '/api/get-users');
@@ -28,12 +30,37 @@ export async function updateUser(id: string, name: string, email: string, age: s
 }
 
 export async function createCart(item_id: number, user_id: string) {
-    return await fetch(`${BASE_URL}/api/carts/create-cart`, {
+    return await fetch(`${BASE_URL}/api/create-cart`, {
         method: 'POST',
         headers: {
             'Content-Type': 'application/json',
-            'Cookie': `user_id=${user_id};`,
+            'Cookie': `uid=${user_id};`,
         },
         body: JSON.stringify({ item_id })
     });
 };
+
+export async function getDetailedCart() {
+    const res = await fetch(`${BASE_URL}/api/get-detailed-cart`, {
+        method: "GET",
+        headers: {
+            "Content-Type": "application/json",
+            "Cookie": `uid=${JSON.parse(cookies().get(AUTH_COOKIE_KEY)?.value!).id};`
+        },
+    });
+
+    const json = await res.json();
+
+    return json;
+}
+
+export async function fetchDataFromApi<T>(
+    url: string,
+    options?: RequestInit | undefined
+): Promise<T> {
+    const res = await fetch(url, options);
+
+    const json = await res.json();
+
+    return json;
+}
